with-sensor/tb-nano-yocto/requirements.txt
certifi==2019.9.11
paho-mqtt==1.5.0
simplejson==3.16.0
yoctopuce==1.10.39263
